fix(cache-server): serve images with the correct content-type header
Until now, we have not bothered setting the `content-type` header when serving images from the cache. This was because I thought we didn't need to -- browsers seemed to handle our images just fine.

It turns out that browsers are pickier with SVG images though, and they will not display an SVG unless it is served with the `image/svg+xml` content-type header.

This PR updates the `cache-server` to use `express`' lovely `res.type()` method, which can automatically determine the correct value to set for the `content-header` cache based on a file extension. See its docs in full here: https://expressjs.com/en/api.html#res.type

A nice side benefit of this change is that we now serve _all_ images with the correct `content-type`, not just SVGs. I don't think this will affect much, but being "more correct" usually doesn't hurt.