fix(errors): fix error summary not working (#5)
Due to a dependency update in NodeCG core itself, the error reporting functionality of Healy (the bit of the UI which enumerates all the individual Validation Errors from the sheet) hasn't been working. This commit get it to work in at least some cases again. There will likely be more work to do to get it working in all cases once again.